refactor: add more subscription debug messages
if a ledger's websocket connection worked but not the subscriptions, the connect promise would never resolve. it was difficult (for all but @justmoon) to debug because the fact that subscribeAccounts was being called after the connect was buried at the end of that massive promise definition. this tries to make it slightly easier to read and adds debug statements to log the RPC requests and responses to make it easier to find RPC errors in the future.